requirements.txt
websocket-client
numpy